Replace in_progress dictionary with global flag
Now that accio uses the quickfix list instead of the location list, it
doesn't make sense to keep a dictionary of all the programs running.  If
even one program is in progress and populating the quickfix list, then
no other program should run.

The most significant effect of this change is no concurrent Accio jobs
will be allowed anymore.

This seems like a step backward, but in truth I need to simplify accio
in order to refine the direction it's headed in.  I'll add support for
concurrent Accio jobs back in later, but for now it will be easier to
focus Accio's development if its underlying model is simple.