Process accio queue on job completion instead of CursorHold
Now that concurrent accio jobs are disallowed, there's no need to wait
for CursorHold to process the queue.  This left the door open for a lot
of problems, like running syntax checkers while the file is in an
invalid state because the user is thinking.

Instead process the queue on job completion, which is the closest Accio
can come to running the command immediately after it was entered by the
user.  This should be good enough for the vast majority of syntax
checkers/linters which finish in sub-second time.