Fix bug: execute compiler program in shell context
Most makeprgs are configured as if they would be executed in a shell
context because this is how Vim executes them.  Unfortunately,
job-control does not execute the program passed to it in any shell
context which may cause errors with some makeprgs.

Accio should forcibly execute the make program in a shell context to
avoid these situations.