Improve multi-line compiler message parsing to be more robust
- Job-control chunks data returned from the compiler program meaning
that Accio only parses chunks of the output message at a time. This
results in a loss of context for multi-line messages which can cause
them to be parsed incorrectly if job-control chunks the output in the
middle of a multi-line message.
    - i.e. a line which is supposed to be a continuation line in a
    larger message is passed to `caddexpr` as the first line of a chunk
    of compiler output.  'errorformat' can only parse the line as a
    continuation line so the line is skipped instead of being parsed as
    a continuation of the previous output chunk

- Solution is to keep track of all lines output from the compiler and
re-parse everything at once instead of in chunks.  No other way to
ensure the chunking effect doesn't corrupt the parsing.