Refactor sign/message handlers. Fix the flickering when clearing signs.
- Clearing signs for a compiler used to cause a flicker because the sign
on a given line would be unplaced, even if there was another sign,
possible with a higher priority, on that line.
- Cleared signs are now replaced with the next best sign/error present
on that line.
- Move signs out of the compiler_task object and instead store all
errors in s:errors_by_line.  The signs for a compiler_task object can be
fetched by combining information in compiler_task.qflist and
s:errors_by_line
- Remove s:accio_messages, all sign errors are now stored in
s:accio_line_errors which has the same indexing as s:accio_messages.
The message text can be fetched from there instead.