Optionally specify statusline format
Adds two optional arguments to accio#statusline(), which are used for
formatting the output. For example, for error counts 0 and 5

    accio#statusline()

produces, respectively

    "Errors: 0" or "Errors: 5"

The format can be specified in the first argument as a printf-style
string:

    accio#statusline("%d errors")

produces, respectively

    "0 errors" or "5 errors"

A second argument can be supplied to specify the output when there are
no errors:

    accio#statusline("%d errors", "No errors")

produces, respectively

    "No errors" or "5 errors"

This last option is particularly useful for passing an empty string if
you don't want to show any indicator when there are no errors.